- Structured streaming example for loading JSON file to a snappy table - Moving structured streaming example under a single folder.
