+++ examples/src/main/scala/org/apache/spark/examples/snappydata/structuredstreaming/SocketSourceExampleWithSnappySink.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018 SnappyData, Inc. All rights reserved.
+ * Copyright (c) 2017-2019 TIBCO Software Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you
  * may not use this file except in compliance with the License. You
@@ -14,38 +14,35 @@
  * permissions and limitations under the License. See accompanying
  * LICENSE file.
  */
+package org.apache.spark.examples.snappydata.structuredstreaming
 
-package org.apache.spark.examples.snappydata
+import scala.reflect.io.Path
 
 import org.apache.log4j.{Level, Logger}
 
+import org.apache.spark.internal.Logging
 import org.apache.spark.sql.streaming.ProcessingTime
 import org.apache.spark.sql.{SnappySession, SparkSession}
-import scala.language.postfixOps
-import scala.reflect.io.Path
 
 /**
- * An example showing usage of structured streaming with SnappyData
- *
- * <p></p>
- * To run the example in local mode go to your SnappyData product distribution
- * directory and type following command on the command prompt
- * <pre>
- * bin/run-example snappydata.StructuredStreamingWithSnappySink
- * </pre>
- * <p>
- * To run this on your local machine, you need to first run a Netcat server <br>
+ * An example showing usage of structured streaming with SnappyData.<br>
+ * To run this on your local machine, you need to first start a Netcat server: <br>
  * `$ nc -lk 9999`
  * <p>
  * Sample input data:
- * {{{
+ * <pre>
  * device1,45
  * device2,67
  * device3,35
- * }}}
+ * </pre>
+ * To run the example in local mode go to your SnappyData product distribution
+ * directory and run the following command:
+ * <pre>
+ * bin/run-example snappydata.structuredstreaming.SocketSourceExampleWithSnappySink
+ * </pre>
  *
  */
-object StructuredStreamingWithSnappySink {
+object SocketSourceExampleWithSnappySink extends Logging {
 
   def main(args: Array[String]) {
     // reducing the log level to minimize the messages on console
@@ -87,14 +84,13 @@ object StructuredStreamingWithSnappySink {
           .writeStream
           .format("snappysink")
           .outputMode("append")
-          .queryName("Devices")
+          .queryName("Devices")  // must be unique across a snappydata cluster
           .trigger(ProcessingTime("1 seconds"))
-          .option("streamQueryId", "Devices") // must be unique across a snappydata cluster
           .option("tableName", "devices")
           .option("checkpointLocation", checkpointDirectory)
           .start()
 
-      streamingQuery.awaitTermination(timeoutMs = 15000)
+      streamingQuery.awaitTermination( 15000)
 
       println("Data loaded in table: ")
       snappy.sql("select * from devices").show()
@@ -111,6 +107,3 @@ object StructuredStreamingWithSnappySink {
   case class DeviceData(device: String, signal: Int)
 
 }
-
-
-
